RD-9224: support for top level methods (#27)
This is a working version, although not sure of all details. I see it
works.

I added the following steps:
* register methods using the existing `recurseFunProto`.
* added a `MethodRefNode` node to implement `IdnExp` when it points to a
method (instead of using `ReadLocalVariable` and the like, I think those
would't work. That node embeds a ref to the top-level function to be
called. It's very similar to a `ClosureNode` except that it doesn't
embed a materialized frame.
* `InvokeNode`: handle both objects `Closure` and `MethodRef` using
`isInstanceOf`. I thought of defining a specific `MethodInvoke` node but
when generating an `InvokeNode`, one doesn't know which kind of object
will be passed as a parameter (e.g. `List.Transform` can receive a
method or a closure).

I enabled a bunch of regression tests (see the code). By doing that I
hit a couple of new bugs. See inline comments.

@Alexzerntev @bgaidioz  There are some questions to be discussed regarding MethodRef and Closure (If they can fit to Closures just with different types of frame)