Added test for RD-9359 (#58)
This is to get a ScalaTest in. It is disabled as a Truffle test because
it should be fixed as part of RD-9029 (support for records with
duplicated field names).

I add a quick fix to the `SemanticAnalyzer`. We forgot to check field
ambiguity for the sugar case of projecting from a list/collection of
records.