RD-9229: data type checks in data exports (#61)
This is the patch to make sure a query isn't run when its result should
be exported to a format that doesn't support the data type (e.g. records
with nested collections when exporting to CSV, but also functions,
packages, when exporting to anything).

I added a test suite to exercise the change, and some existing tests
were enabled.