Fixed TextOutputTest assertion (#101)
That test is breaking against the Scala executor that reports a slightly
different error message. Maybe we can be a little relaxed in the
assertion to make both pass? (I tried that one in Scala and it passed.)

The scala failure has both 'unsupported type' and 'expected either
common string or option(common string) but got common int'.