Remove deprecated code. (#115)
This PR removes leftover/dead code related to RQL. This includes classes, config settings, etc.
The most visible change is that rql""" is now snapi""".
It also applies some renames to classes that were conflicting with other projects, e.g. Service becomes RawService.
ProgramSettings is now gone; this was a complicated mechanism used to pass settings. This is no longer needed since the "options" are now available in the ProgramEnvironment.
The JIRACrashReporter is also removed from this codebase.