Persistent pool optimized for FDW (#466)
Implements a connection pool that is designed specifically for the needs of FDW and Steampipe.
This connection pool tries to keep connections active/alive for as long as possible; this ensures that Steampipe caches, which are per connection, stay alive for as long as possible.
The connection pool has a total global limit on the number of connections. If a new connection needs to be established, it will synchronously try to remove the oldest connection that isn't currently borrowed by anyone.
In addition, it keeps a limit on the number of active connections to a single database - this is done to prevent a single "user" from taking over all available slots.
Finally, it has a background thread that checks the health state of connections in the pool; this picks up one connection at a time (that is not used recently), and uses the JDBC isValid(...) method to check its validity.